requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.4.0
certifi==2021.10.8
charset-normalizer==2.0.10
frozenlist==1.3.0
idna==3.3
multidict==5.2.0
nextcord==2.0.0a6
python-dotenv==0.19.2
requests==2.27.1
urllib3==1.26.8
yarl==1.7.2
